requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
cryptography==37.0.2
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
orjson==3.7.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pypika-tortoise==0.1.5
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tortoise-orm==0.19.1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
